auth_webapp/requirements.txt
django==3.2.2
python-dotenv
psycopg2
git+https://github.com/mailchimp/mailchimp-marketing-python.git
pyjwt